requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
SQLAlchemy==1.4.49
yfinance
apache-airflow==2.8.1